ggspider not spawning fix attempt
Mesh_bounds?
Couldn't reproduce the issue on my end with mesh_bounds interacting with enemy mesh_bounds I reckon it's not the source of the issue. reducing ggspider mesh_bound below spawner mesh_bound just in case.

Energy priority disabled?
Could reproduce the issue by setting energy priority to low so the spawner won't spawn the unit. This may not be the real source of the issue.
Disabling ORDER_ATTACK fixes it by not being able to touch energy priority to this unit even thought it doesn't consume energy.
Although Player now cannot fire the spawner but using attack command so the unit won't spawn ggspider if no enemy is on the planet.